Moved SequenceFailureBehavior parameter to JobbyServerSettings from ProgresqlStorageSettings.
