+++ src/Jobby.Core/Interfaces/Configuration/ICommonInfrastructure.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using Jobby.Core.Models;
+using Microsoft.Extensions.Logging;
 
 namespace Jobby.Core.Interfaces.Configuration;
 
@@ -7,4 +8,5 @@ public interface ICommonInfrastructure
     ILoggerFactory LoggerFactory { get; }
     IJobParamSerializer Serializer { get; }
     IGuidGenerator GuidGenerator { get; }
+    JobbyServerSettings ServerSettings { get; }
 }
