+++ src/Jobby.Postgres/Commands/BulkInsertJobsCommand.cs
@@ -10,11 +10,11 @@ internal class BulkInsertJobsCommand
     private readonly string _commandText;
     private readonly string _commandWithSequenceIdText;
 
-    public BulkInsertJobsCommand(NpgsqlDataSource dataSource, PostgresqlStorageSettings settings)
+    public BulkInsertJobsCommand(NpgsqlDataSource dataSource, PostgresqlStorageSettings settings, JobbyServerSettings serverSettings)
     {
         _dataSource = dataSource;
 
-        var blockSequenceOnFailure = settings.SequenceFailureBehavior == SequenceFailureBehavior.Block;
+        var blockSequenceOnFailure = serverSettings.SequenceFailureBehavior == SequenceFailureBehavior.Block;
 
         _commandText = @$"
             INSERT INTO {TableName.Jobs(settings)} (
