+++ src/Jobby.Postgres/ConfigurationExtensions/IPostgresqlStorageConfigurable.cs
@@ -1,5 +1,4 @@
-﻿using Jobby.Postgres;
-using Npgsql;
+﻿using Npgsql;
 
 namespace Jobby.Postgres.ConfigurationExtensions;
 
@@ -8,5 +7,4 @@ public interface IPostgresqlStorageConfigurable
     IPostgresqlStorageConfigurable UseDataSource(NpgsqlDataSource dataSource);
     IPostgresqlStorageConfigurable UseSchemaName(string schemaName);
     IPostgresqlStorageConfigurable UseTablesPrefix(string tablesPrefix);
-    IPostgresqlStorageConfigurable UseSequenceFailureBehavior(SequenceFailureBehavior behavior);
 }
