+++ src/Jobby.Postgres/ConfigurationExtensions/PostgresqlStorageBuilder.cs
@@ -1,5 +1,6 @@
 ﻿using Jobby.Core.Exceptions;
 using Jobby.Core.Interfaces.Configuration;
+using Jobby.Core.Models;
 using Microsoft.Extensions.Logging;
 using Npgsql;
 
@@ -21,8 +22,7 @@ internal class PostgresqlStorageBuilder : IPostgresqlStorageConfigurable
         _settings = new PostgresqlStorageSettings
         {
             SchemaName = schemaName,
-            TablesPrefix = _settings.TablesPrefix,
-            SequenceFailureBehavior = _settings.SequenceFailureBehavior
+            TablesPrefix = _settings.TablesPrefix
         };
         return this;
     }
@@ -32,30 +32,18 @@ internal class PostgresqlStorageBuilder : IPostgresqlStorageConfigurable
         _settings = new PostgresqlStorageSettings
         {
             SchemaName = _settings.SchemaName,
-            TablesPrefix = tablesPrefix,
-            SequenceFailureBehavior = _settings.SequenceFailureBehavior
+            TablesPrefix = tablesPrefix
         };
         return this;
     }
 
-    public IPostgresqlStorageConfigurable UseSequenceFailureBehavior(SequenceFailureBehavior behavior)
+    internal PostgresqlJobbyStorage BuildStorage(JobbyServerSettings serverSettings)
     {
-        _settings = new PostgresqlStorageSettings
-        {
-            SchemaName = _settings.SchemaName,
-            TablesPrefix = _settings.TablesPrefix,
-            SequenceFailureBehavior = behavior
-        };
-        return this;
-    }
-
-    internal PostgresqlJobbyStorage BuildStorage()
-    {
-        if (_dataSource == null) 
+        if (_dataSource == null)
         {
             throw new InvalidBuilderConfigException("DataSource is not configured for PostgresStorage. UseDataSource method should be called");
         }
-        return new PostgresqlJobbyStorage(_dataSource, _settings);
+        return new PostgresqlJobbyStorage(_dataSource, _settings, serverSettings);
     }
 
     internal PostgresqlJobbyStorageMigrator BuildMigrator(ICommonInfrastructure commonInfra)
