+++ src/Jobby.Postgres/PostgresqlStorageSettings.cs
@@ -4,5 +4,4 @@ internal class PostgresqlStorageSettings
 {
     public string SchemaName { get; init; } = string.Empty;
     public string TablesPrefix { get; init; } = "jobby_";
-    public SequenceFailureBehavior SequenceFailureBehavior { get; init; } = SequenceFailureBehavior.Block;
 }
